add expand of one_click (#72)
* add one_click for expand

* add expand

* add opt for cdn

* fix config.ini

* test ci. (#74)

* update install deps, do not check sudo permission.

* update pack.sh

* remove extra p2p section

* resolve conflict

* try to fix ci

* fix fisco-bcos not exist. (#75)

* update install deps, do not check sudo permission.

* update pack.sh

* fix fisco-bcos not exist.

* fix ci at ubuntu at 18.10

* fix issues at codefactor

* fix exit at centos6

* update cts (#77)

* update cts

* update one_click_expand

* update readme for meta

* fix readme of grammarly

* add contribute doc and update readme doc (#78)

* update readme (#76)

* add contribute doc and update readme doc

* fix misspelling Continuous

* fix using wrong ca.key

+++ one_click_generator.sh
@@ -6,27 +6,29 @@
 # @ file     : one_click_generator.sh
 # @ date     : 2019
 
+branch_name=$(git symbolic-ref --short -q HEAD)
+
 SHELL_FOLDER=$(
     cd $(dirname $0)
     pwd
 )
 
-output_dir=$1
+output_dir=$2
 dir_name=()
+group_id=
 EXIT_CODE=1
 
 help() {
-    echo "View at https://fisco-bcos-documentation.readthedocs.io/zh_CN/release-2.0.0/docs/enterprise_tools/enterprise_quick_start.html"
+    echo "bash one_click_generator.sh -b ./tmp_onc_click"
+    echo "bash one_click_generator.sh -e ./tmp_onc_click_expand"
+    echo "View at https://fisco-bcos-documentation.readthedocs.io/zh_CN/release-2.0/docs/enterprise_tools/enterprise_quick_start.html"
 }
 
 check_result() {
     ret=$?
     if [ $ret -ne 0 ]; then
-        LOG_ERROR "FAILED execution of command: ${command}"
-        # clear_cache
-        exit 1
-    else
-        LOG_INFO "SUCCESS execution of command: ${command}"
+        LOG_ERROR "STATUS CHECK FAILED"
+        exit $EXIT_CODE
     fi
 }
 
@@ -67,9 +69,49 @@ dir_must_exists() {
 }
 
 dir_must_not_exists() {
-    if [ -e "$1" ]; then
+    if [ -d "$1" ]; then
         echo "$1 DIR exists, please clean old DIR!"
         exit $EXIT_CODE
+    fi
+}
+
+# check_rename() {
+#     if [ -d "$1" ]; then
+#         echo "$1 exists, Want to rename it?(y/n)"
+#         read status
+#         if [ "${status}" == "y" ]; then
+#             LOG_INFO "input name at local path"
+#             read name
+#             mv $1 ${name}
+#         else
+#             LOG_ERROR "$1 exist, break!"
+#             exit $EXIT_CODE
+#         fi
+#     fi
+# }
+
+check_generator_status () {
+    file_must_not_exists ${SHELL_FOLDER}/meta/ca.*
+    file_must_not_exists ${SHELL_FOLDER}/meta/agency.*
+    file_must_not_exists ${SHELL_FOLDER}/meta/cert_*
+    dir_must_not_exists ${SHELL_FOLDER}/meta/node_*
+    if [ ! -f "${SHELL_FOLDER}/meta/fisco-bcos" ]; then
+        LOG_INFO "doanloading fisco-bcos..."
+        ./generator --download_fisco ./meta
+        check_result
+    fi
+}
+
+get_time_stamp() {
+    current=$(date "+%Y-%m-%d %H:%M:%S")
+    timeStamp=$(date -d "$current" +%s)
+    echo ${timeStamp}
+}
+
+check_rename() {
+    if [ -d "$1" ]; then
+        echo "$1 exists, rename it to $1_$(get_time_stamp)"
+        mv $1 $1_$(get_time_stamp)
     fi
 }
 
@@ -80,52 +122,48 @@ check_node_ini() {
         if [ -d ${dir} ]; then
             if [ ! -f "${dir}/node_deployment.ini" ]; then
                 LOG_ERROR "${dir}/node_deployment.ini not exist!"
-                exit 1
+                exit $EXIT_CODE
+            fi
+            if [ -d "${dir}/generator-agency" ]; then
+                LOG_ERROR "${dir}/generator-agency exist!"
+                exit $EXIT_CODE
             fi
             LOG_INFO "try to use ${dir}/node_deployment.ini"
-            if [ -d "${dir}/generator" ]; then
-                LOG_ERROR "${dir}/generator existed, please delete it!"
-                exit 1
-            fi
             dir_name[i]=${dir}
             i=${i}+1
+            group_id=$(< ${dir}/node_deployment.ini grep group_id= | tr -cd '0-9')
         fi
     done
 
 }
 
-run_install() {
-    cd ${SHELL_FOLDER}/
-    bash ./scripts/install.sh
-}
 
 init_chain() {
     cd ${SHELL_FOLDER}/
-    ./generator --generate_chain_certificate ./dir_chain_ca
+    ./generator --generate_chain_certificate ./.dir_chain_ca
     check_result
-    if [ ! -f "${SHELL_FOLDER}/meta/fisco-bcos" ]; then
-        LOG_INFO "doanloading fisco-bcos..."
-        ./generator --download_fisco ./meta
-        check_result
-    fi
-}
-
-init_agency_with_ke() {
-    echo "no key"
-
+    cp ./.dir_chain_ca/* ${output_dir}/
+    rm -rf ./.dir_chain_ca
 }
 
 init_agency() {
     for agency in ${dir_name[*]}; do
         cd ${SHELL_FOLDER}
-        file_must_not_exists ${agency}/agency.crt
-        file_must_not_exists ${agency}/agency.key
-        git clone https://github.com/FISCO-BCOS/generator.git ${agency}/generator-agency
-        cp ${SHELL_FOLDER}//meta/fisco-bcos ${agency}/generator-agency/meta/fisco-bcos
-        ./generator --generate_agency_certificate ${agency} ./dir_chain_ca agency_cert
-        check_result
-        cp ${agency}/agency_cert/* ./${agency}/generator-agency/meta
-        cp ${agency}/agency_cert/* ${SHELL_FOLDER}/meta/
+        if [ ! -d ${agency}/generator-agency ]; then
+            git clone https://github.com/FISCO-BCOS/generator.git ${agency}/generator-agency -b ${branch_name}
+            cp ${SHELL_FOLDER}/meta/fisco-bcos ${agency}/generator-agency/meta/fisco-bcos
+        fi
+        if [ -f "${agency}/agency_cert/agency.crt" ]; then
+            if [ -f "${agency}/agency_cert/agency.crt" ]; then
+                cp ${agency}/agency_cert/* ./${agency}/generator-agency/meta
+            fi
+        else
+            file_must_not_exists ${agency}/agency_cert/agency.crt
+            ./generator --generate_agency_certificate ${agency} ${output_dir}/ agency_cert
+            check_result
+            cp ${agency}/agency_cert/* ./${agency}/generator-agency/meta
+            # cp ${agency}/agency_cert/* ${SHELL_FOLDER}/meta/
+        fi
     done
 }
 
@@ -137,6 +175,7 @@ init_node_cert() {
         cd ${agency}/generator-agency
         cp -r ../node_deployment.ini ./conf/
         LOG_INFO "${agency} generate node now..."
+        check_rename ./node_send
         ./generator --generate_all_certificates ./node_send
         check_result
         cp ./node_send/cert* ${SHELL_FOLDER}/meta
@@ -146,7 +185,11 @@ init_node_cert() {
     if [ -f "${SHELL_FOLDER}/meta/peers.txt" ]; then
         rm ${SHELL_FOLDER}/meta/peers.txt
     fi
+    if [ -f "${output_dir}/peers.txt" ]; then
+        cat ${output_dir}/peers.txt >>${SHELL_FOLDER}/meta/peersALL.txt
+    fi
     sort -n ${SHELL_FOLDER}/meta/peersALL.txt | uniq >${SHELL_FOLDER}/meta/peers.txt
+    cp ${SHELL_FOLDER}/meta/peers.txt ${output_dir}/
 }
 
 init_genesis() {
@@ -157,7 +200,7 @@ init_genesis() {
     fi
     cat <<EOF >>"./conf/group_genesis.ini"
 [group]
-group_id=1
+group_id=${group_id}
 
 [nodes]
 EOF
@@ -167,48 +210,131 @@ EOF
         i=$((i + 1))
     done <./meta/peers.txt
 
-    ./generator --create_group_genesis ./group
+    ./generator --create_group_genesis ./.group
+    cp ./.group/group.*.genesis ${output_dir}/
     check_result
     for agency in ${dir_name[*]}; do
-        cp ./group/group.1.genesis ${agency}/generator-agency/meta/
+        cp ./.group/group.*.genesis ${SHELL_FOLDER}/${agency}/generator-agency/meta/
     done
-
+    rm -rf ./.group
+    rm ${SHELL_FOLDER}/meta/cert_*
+    rm ${SHELL_FOLDER}/meta/group.*
+    rm -rf ${SHELL_FOLDER}/meta/node_*
 }
 
 generate_node() {
     cd ${SHELL_FOLDER}
     for agency in ${dir_name[*]}; do
         cd ${agency}/generator-agency/
+        check_rename ./node
         ./generator --build_install_package ${SHELL_FOLDER}/meta/peers.txt ./node
         check_result
+        check_rename ./sdk
         ./generator --get_sdk_file ./sdk
         check_result
-        cp -rf ./node ${SHELL_FOLDER}/${agency}/node
-        # cp -r ${SHELL_FOLDER}/console ${agency}/console
-        cp -rf ./sdk/ ${SHELL_FOLDER}/${agency}/sdk
+        check_rename ${SHELL_FOLDER}/${agency}/node
+        check_rename ${SHELL_FOLDER}/${agency}/sdk
+        mv ./node ${SHELL_FOLDER}/${agency}/node
+        mv ./sdk/ ${SHELL_FOLDER}/${agency}/sdk
         cd ${SHELL_FOLDER}
     done
+    if [ -f "${SHELL_FOLDER}/meta/peers*" ]; then
+        rm ${SHELL_FOLDER}/meta/peers*
+    fi
+    if [ -f "${SHELL_FOLDER}/meta/cert_*" ]; then
+        rm ${SHELL_FOLDER}/meta/cert_*
+    fi
 }
 
-# download_console() {
-#     cd ${SHELL_FOLDER}
-#     ./generator --download_console ./
-# }
-
-main() {
-    run_install
+build_init() {
+    # run_install
+    check_generator_status
     init_chain
-    check_node_ini ${output_dir}
+    check_node_ini
     init_agency
     init_node_cert
     init_genesis
-    # download_console
     generate_node
+}
+
+chain_must_exist() {
+    file_must_exists ${output_dir}/ca.crt
+    file_must_exists ${output_dir}/ca.key
+    file_must_exists ${output_dir}/peers.txt
+
+}
+
+copy_genesis() {
+    for agency in ${dir_name[*]}; do
+        cp ${output_dir}/group.${group_id}.genesis ${SHELL_FOLDER}/${agency}/generator-agency/meta/
+    done
+}
+
+add_peers_expand_node() {
+    cat ${output_dir}/peers.txt >>${SHELL_FOLDER}/meta/peers.txt
+    sort -n ${SHELL_FOLDER}/meta/peers.txt | uniq
+}
+
+expand_init() {
+    check_generator_status
+    chain_must_exist
+    expand_node_ini
+    init_agency
+    init_node_cert
+    copy_genesis
+    add_peers_expand_node
+    generate_node
+}
+
+expand_node_ini() {
+    i=0
+    for dir in ${output_dir}/*; do
+        [[ -e ${dir} ]] || break
+        if [ -d ${dir} ]; then
+            if [ ! -f "${dir}/node_deployment.ini" ]; then
+                LOG_ERROR "${dir}/node_deployment.ini not exist!"
+                exit $EXIT_CODE
+            fi
+            if [ -d "${dir}/generator-agency" ]; then
+                LOG_ERROR "${dir}/generator-agency exist!"
+                exit $EXIT_CODE
+            fi
+            # group_id=$(cat ${dir}/node_deployment.ini | grep group_id= | tr '\r' ' ' | sed s/ //g | sed s/group_id=//g)
+            group_id=$(< ${dir}/node_deployment.ini grep group_id= | tr -cd '0-9')
+            file_must_exists ${output_dir}/group.${group_id}.genesis
+            LOG_INFO "Need group.${group_id}.genesis in ${output_dir}"
+            LOG_INFO "try to use ${dir}/node_deployment.ini"
+            dir_name[i]=${dir}
+            i=${i}+1
+        fi
+    done
 }
 
 if [ -z "$1" ]; then
     LOG_ERROR "not input found!"
     help
-else
-    main $1
 fi
+
+case "$1" in
+-b)
+    build_init $2
+    ;;
+-build)
+    build_init $2
+    ;;
+-e)
+    expand_init $2
+    ;;
+-expand)
+    expand_init $2
+    ;;
+help)
+    help
+    ;;
+*)
+    help
+    ;;
+esac
+
+check_result
+echo -e "\033[32m run one_click_generator successful!\033[0m"
